REFACTOR :: 코드 컨벤션 맞게 수정
