+++ src/app/layout.tsx
@@ -1,7 +1,7 @@
 import "@/styles/global.css";
 import { ReactNode } from "react";
 
-interface PropsInterface {
+interface PropsType {
   children: ReactNode;
 }
 
@@ -10,7 +10,7 @@ export const metadata = {
   description: "부산소마고의 입학을 책임지는 마루입니다.",
 };
 
-const RootLayout = ({ children }: PropsInterface) => {
+const RootLayout = ({ children }: PropsType) => {
   return (
     <html lang="kr">
       <body>{children}</body>
