+++ src/components/Common/Button/index.ts
@@ -27,10 +27,9 @@ export const getButtonStyle = {
   `,
   QUATERNARY: css`
     background-color: ${color.white};
-    color: ${color.gray600};
+    color: ${color.gray900};
     border: none;
     &:hover {
-      color: ${color.gray900};
       background-color: ${color.gray200};
     }
   `,
@@ -47,10 +46,11 @@ export const getButtonStyle = {
     border-radius: 0;
     padding: 0;
     color: ${color.gray500};
-    border-bottom: 1px solid ${color.gray500};
+    text-decoration-line: underline;
+    text-decoration-color: ${color.gray500};
     &:hover {
       color: ${color.maruDefault};
-      border-bottom: 1px solid ${color.maruDefault};
+      text-decoration-color: ${color.maruDefault};
     }
   `,
 };
