+++ src/layouts/AppLayout.tsx
@@ -2,12 +2,12 @@ import { color } from "@/styles/color";
 import { ReactNode } from "react";
 import styled from "styled-components";
 
-interface PropsInterface {
+interface PropsType {
   children: ReactNode;
   height?: string;
 }
 
-const AppLayout = ({ children, height }: PropsInterface) => {
+const AppLayout = ({ children, height }: PropsType) => {
   return <StyledAppLayout style={{ height }}>{children}</StyledAppLayout>;
 };
